small miner refactor/move
